Move VerifyStatus enum to lib
